cosmic_freedesktop_icons

Struct LookupBuilder

pub struct LookupBuilder<'a> { /* private fields */ }

The lookup builder struct, holding all the lookup query parameters.

Implementations

impl<'a> LookupBuilder<'a>

pub fn with_size(self, size: u16) -> Self

Restrict the lookup to the given icon size.

Example

use cosmic_freedesktop_icons::lookup;

let icon = lookup("firefox")
    .with_size(48)
    .find();

pub fn with_scale(self, scale: u16) -> Self

Restrict the lookup to the given scale.

Example

use cosmic_freedesktop_icons::lookup;

let icon = lookup("firefox")
    .with_scale(2)
    .find();

pub fn with_theme<'b: 'a>(self, theme: &'b str) -> Self

Add the given theme to the current lookup :

Example

use cosmic_freedesktop_icons::lookup;

let icon = lookup("firefox")
    .with_theme("Papirus")
    .find();

pub fn with_cache(self) -> Self

Store the result of the lookup in cache, subsequent lookup will first try to get the cached icon. This can drastically increase lookup performances for application that repeat the same lookups, an application launcher for instance.

Example

use cosmic_freedesktop_icons::lookup;

let icon = lookup("firefox")
    .with_scale(2)
    .with_cache()
    .find();

pub fn force_svg(self) -> Self

By default [find] will prioritize Png over Svg icon. Use this if you need to prioritize Svg icons. This could be useful if you need a modifiable icon, to match a user theme for instance.

Example

use cosmic_freedesktop_icons::lookup;

let icon = lookup("firefox")
    .force_svg()
    .find();

pub fn find(self) -> Option<PathBuf>

Execute the current lookup if no icon is found in the current theme fallback to /usr/share/icons/hicolor theme and then to /usr/share/pixmaps.

pub fn cache_clear(&mut self)

pub fn cache_reset_none(&mut self)